+++ fairing.go
@@ -6,8 +6,8 @@ import (
 	"github.com/dannypsnl/rocket/response"
 )
 
-// fairingInterface specify the method that fairing could implement
-type fairingInterface interface {
+// Fairing specify the method that fairing could implement
+type Fairing interface {
 	OnRequest(*http.Request) *http.Request
 	OnResponse(*response.Response) *response.Response
 	// OnLaunch would let you could get the metadata of rocket server
@@ -17,14 +17,14 @@ type fairingInterface interface {
 	OnLaunch(*Rocket)
 }
 
-// Fairing provides default implement for your fairing by embedded it into your fairing type,
+// DefaultFairing provides default implement for your fairing by embedded it into your fairing type,
 // Embedded is a good practice because fairing is designed to allowed partial implement interface
-type Fairing struct{}
+type DefaultFairing struct{}
 
-func (f Fairing) OnRequest(req *http.Request) *http.Request {
+func (f DefaultFairing) OnRequest(req *http.Request) *http.Request {
 	return req
 }
-func (f Fairing) OnResponse(resp *response.Response) *response.Response {
+func (f DefaultFairing) OnResponse(resp *response.Response) *response.Response {
 	return resp
 }
-func (f Fairing) OnLaunch(*Rocket) {}
+func (f DefaultFairing) OnLaunch(*Rocket) {}